Add ansible/shell.nix. (With story in comments.)
This no longer seems as necessary as it did earlier, but it may be handy to keep around.

Basically, I was trying to follow docs for the Ansible-GCE integration, which
(apparently, like many Ansible plugins) is written in Python. So then I needed
extra Python libraries. And this addressed that problem.

But then I started going through configuration with Ansible-GCE.  It
requires extra a credentials file (*which I happen to have lying around --
but other folks won't, and it's annoying to get*), and I also found
some important bits felt under-documented (re: mapping between
ansible metadata and google metadata). Maybe solvable, but...

I threw out the plugin and wrote `bin/update-gcloud`.  It still requires
some setup (eg `gcloud auth login`), but I think it's simpler arrangement
overall.

Still -- based on the observation that Python modules are common in Ansible,
it seems likely that we'll need them for something. So having this example
reassures me that I'll be able to use similar plugins (when/if needed).